Update
Updates:
-----------
- Added Datastore Manager to store user info beyond the Activity lifecycle
- Added caching crypto bubbles after the app is launched so that the API doesn't get spammed and cause limit/response bug for every screen switch
- Added 'add to favorites' functionality with a list in datastore
- Added a favorites list screen (Needs work)
- Added many features to API's
- Fixed small bug with coingecko data class
- Many small changes/fixes

TODO:
--------
- Work on Favorites list screen
- Add suggested searching (Maybe display a list after the first 2 letters are present? - better way to know what coins are available to search)
- Add indicators/price alerts to the chart
- Add notes section
- Add TA/indicators section

+++ app/src/main/java/com/willkopec/whalert/model/coingecko/CryptoItem.kt
@@ -26,5 +26,5 @@ data class CryptoItem(
     val roi: Roi?,
     val symbol: String,
     val total_supply: Double,
-    val total_volume: Long
+    val total_volume: Double
 )
